+++ engine/inc/libCoreCommon/actor_id_converter.h
@@ -10,6 +10,6 @@ namespace core
 		CActorIDConverter() { }
 		virtual ~CActorIDConverter() { }
 
-		virtual uint16_t convertToServiceID(uint64_t nActorID) = 0;
+		virtual uint32_t convertToServiceID(uint64_t nActorID) = 0;
 	};
 }
