+++ engine/inc/libCoreCommon/base_app.h
@@ -35,7 +35,7 @@ namespace core
 		nStartTime 第一次触发定时器的时间
 		nIntervalTime 第一次触发定时器后接下来定时器触发的间隔时间，如果该值是0就表示这个定时器只触发一次
 		*/
-		void						registerTicker(uint8_t nType, uint16_t nFromServiceID, uint64_t nFromActorID, CTicker* pTicker, uint64_t nStartTime, uint64_t nIntervalTime, uint64_t nContext);
+		void						registerTicker(uint8_t nType, uint32_t nFromServiceID, uint64_t nFromActorID, CTicker* pTicker, uint64_t nStartTime, uint64_t nIntervalTime, uint64_t nContext);
 		/**
 		@brief: 反注册定时器
 		*/
@@ -63,7 +63,7 @@ namespace core
 		/*
 		@brief: 根据服务ID获取服务
 		*/
-		CServiceBase*				getServiceBase(uint16_t nServiceID) const;
+		CServiceBase*				getServiceBase(uint32_t nServiceID) const;
 		/**
 		@brief: 获取本节点基本信息
 		*/
@@ -71,11 +71,11 @@ namespace core
 		/**
 		@brief: 获取本节点基本信息
 		*/
-		uint16_t					getNodeID() const;
+		uint32_t					getNodeID() const;
 		/**
 		@brief: 根据节点名字获取节点id
 		*/
-		uint16_t					getServiceID(const std::string& szName) const;
+		uint32_t					getServiceID(const std::string& szName) const;
 		/*
 		@brief: 获取配置文件名
 		*/
