+++ engine/src/libCoreCommon/core_connection_mgr.h
@@ -28,7 +28,7 @@ namespace core
 		bool				init(uint32_t nMaxConnectionCount);
 		bool				connect(const std::string& szHost, uint16_t nPort, uint32_t nType, const std::string& szContext, uint32_t nSendBufferSize, uint32_t nRecvBufferSize, MessageParser messageParser);
 		bool				listen(const std::string& szHost, uint16_t nPort, uint32_t nType, const std::string& szContext, uint32_t nSendBufferSize, uint32_t nRecvBufferSize, MessageParser messageParser);
-		void				update(int64_t nTime);
+		void				update(uint32_t nTime);
 
 		void				broadcast(uint32_t nType, uint8_t nMessageType, const void* pData, uint16_t nSize, uint64_t* pExcludeID, uint16_t nExcludeIDCount);
